Update README
Reword the opening section of the README for clarity. It took me a few clicks to realise that the second link is the super-important guide to how to actually run the docker image (this page might profitably be linked in the docker image readme too). Hopefully it's more obvious now.